backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.0"
description = ""
authors = ["Aleksey Semenov <alexsemenov1610@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.110.0"}
pyjwt = "^2.8.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
bcrypt = "^4.1.2"
mimesis = "^16.0.0"
tqdm = "^4.66.2"
aio-pika = "^9.4.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
